DBGM Role Opting
New command allows users to opt-in/out of specified roles by typing (prefix)iam <rolealias>.

The role aliases are specified by an admin (i.e. someone of adequate permissions) who must assign aliases to roles. Roles cannot be opted into that do not have an alias (so users can't opt-in to admin roles... Unless the admin creates an alias for that).