Foundation GuildMessage Update
Changed MessageReceivedEvent to GuildMessageReceivedEvent... This broke everything, so I'm going to be updating each of my other bots independently.

The reason for this change is because I had been using the MessageReceivedEvent as though it were a GuildMessageReceivedEvent, so there was a lot of redundency being called that I didn't want to have break anything potentially... So I changed it.

I also changed the Respond class to only take TextChannels as an argument. Before they took MessageReceivedEvents too, but I got rid of those...