DBGM Purge Overhaul
Optimised the iterative process for reviewing message history.

Built in a sensor for whether a message is too old to be deleted in the SelfPurgeList

Used Foundation's getMemberByMention utility.

Added the ability for me to, sometime in the future, implement the ability for someone to mention multiple users to purge... I haven't actually gotten that working yet.