Foundation - Bot Mention Commands
Fixed a problem where if you nickname the bot, then it would not respond to any mention of it. The reason for this being is that if a person has a nickname, then their mention in raw text includes an !... This messed up my regex and how I determined whether the bot's mention preceded the message, so I had to redo that stuff with more regex.