DBGM Stream Watch
Added the capability (theoretically; I haven't actually tested to see if it works yet) for the bot to send a message within a specified (the default if unspecified) channel when a user starts streaming.

Only users specified by the 'stream' command will cause the bot to trigger a server notification.

The code relies on the user's Presence changing. If the user's Presence does not change to reflect that "So-and-so is Streaming Such-and-such", then the bot will not know that they are streaming.