DND - DiceRoller
Oohh boy. Added the Roll command. The only function of this command is to 'init' a channel. When a channel is successfully inited, then the bot will delete all messages in that channel that it can and begin to monitor the channel for dice rolls (e.g. d20, d6+1, d100-2, etc).

Any message sent into that channel will be deleted. Any valid dice roll will be added to the history of dice rolls. The channel will be saved and reloaded when the bot shuts down and starts up. The bot can be disabled from listening to that channel if someone types 'uninit' instead of a dice roll.

This was a mess. I'll probably revisit this later, but it's working.