Foundation - Command/EventListener AutoUnload
Do I get points for a technically fully descriptive, really long commit title? No?

Just made it so that when the unloadAll method unloads each command, it also checks to see if it is an EventListener and removes it from the JDA if it is so that there isn't a double-save after unloadAll finishes.