Stopwatch
Added the stopwatch command to DBGM and reorganized somethings in Foundation to make things a bit easier.

The stopwatch basically just saves the time you started it at, and whenever you ask, it will tell you exactly how long it's been since you started the stopwatch. Mainly just for me waiting on crap to finish in my games and being too lazy to log in and check when they're done.